removing factorytest
