+++ server/app/services/ProgramBlockValidation.java
@@ -1,5 +1,7 @@
 package services;
 
+import static com.google.common.base.Preconditions.checkNotNull;
+
 import java.util.Optional;
 import models.Version;
 import services.program.BlockDefinition;
