+++ server/test/services/ProgramBlockValidationTest.java
@@ -2,12 +2,15 @@ package services;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import models.Question;
+import models.Version;
 import org.junit.Before;
 import org.junit.Test;
 import repository.ResetPostgres;
 import repository.VersionRepository;
 import services.ProgramBlockValidation.AddQuestionResult;
 import services.program.ProgramDefinition;
+import services.question.QuestionService;
 import services.question.types.QuestionDefinition;
 import support.ProgramBuilder;
 
@@ -15,13 +18,38 @@ public class ProgramBlockValidationTest extends ResetPostgres {
 
   private ProgramBlockValidation programBlockValidation;
   private VersionRepository versionRepository;
+  private QuestionService questionService;
+  private Question questionForTombstone;
 
   @Before
-  public void setProgramBlockValidation() {
-    versionRepository = instanceOf(VersionRepository.class);
+  public void setProgramBlockValidation()
+      throws services.question.exceptions.QuestionNotFoundException {
+    questionService = instanceOf(services.question.QuestionService.class);
+    versionRepository = instanceOf(repository.VersionRepository.class);
+    Version version = versionRepository.getDraftVersion();
+    String tombstonedQuestionOneName = "tombstoneOne";
+    questionForTombstone = resourceCreator.insertQuestion(tombstonedQuestionOneName);
+    version.addQuestion(questionForTombstone);
+    version.addTombstoneForQuestion(questionForTombstone);
+    version.save();
     ProgramBlockValidationFactory programBlockValidationFactory =
         new ProgramBlockValidationFactory(versionRepository);
     programBlockValidation = programBlockValidationFactory.create();
+  }
+
+  @Test
+  public void canAddQuestions_cantAddArchivedQuestion() throws Exception {
+    ProgramDefinition program =
+        ProgramBuilder.newDraftProgram("program1")
+            .withBlock()
+            .withRequiredQuestionDefinition(questionForTombstone.getQuestionDefinition())
+            .buildDefinition();
+    assertThat(
+            programBlockValidation.canAddQuestion(
+                program,
+                program.getLastBlockDefinition(),
+                questionForTombstone.getQuestionDefinition()))
+        .isEqualTo(services.ProgramBlockValidation.AddQuestionResult.QUESTION_TOMBSTONED);
   }
 
   @Test
